🚚 api: move session errors to src directory
