+++ api/lib/domain/errors.js
@@ -755,39 +755,9 @@ class PasswordResetDemandNotFoundError extends DomainError {
   }
 }
 
-class SessionAlreadyFinalizedError extends DomainError {
-  constructor(message = 'Cannot finalize session more than once.') {
-    super(message);
-    this.code = 'SESSION_ALREADY_FINALIZED';
-  }
-}
-
+// FIXME: used ?
 class SessionWithIdAndInformationOnMassImportError extends DomainError {
   constructor(message = 'Merci de ne pas renseigner les informations de session') {
-    super(message);
-  }
-}
-
-class SessionWithoutStartedCertificationError extends DomainError {
-  constructor(message = "This session hasn't started, you can't finalise it. However, you can delete it.") {
-    super(message);
-    this.code = 'SESSION_WITHOUT_STARTED_CERTIFICATION';
-  }
-}
-
-class SessionWithAbortReasonOnCompletedCertificationCourseError extends DomainError {
-  constructor(
-    message = 'The field "Reason for abandonment" has been filled in for a candidate who has finished their certification exam in between. The session therefore can\'t be finalised. Please refresh the page before finalising.',
-  ) {
-    super(message);
-    this.code = 'SESSION_WITH_ABORT_REASON_ON_COMPLETED_CERTIFICATION_COURSE';
-  }
-}
-
-class SessionWithMissingAbortReasonError extends DomainError {
-  constructor(
-    message = "Une ou plusieurs certifications non terminées n'ont pas de “Raison de l’abandon” renseignées. La session ne peut donc pas être finalisée.",
-  ) {
     super(message);
   }
 }
@@ -1281,13 +1251,9 @@ export {
   SendingEmailToInvalidEmailAddressError,
   SendingEmailToRefererError,
   SendingEmailToResultRecipientError,
-  SessionAlreadyFinalizedError,
   SessionAlreadyPublishedError,
   SessionNotAccessible,
-  SessionWithAbortReasonOnCompletedCertificationCourseError,
   SessionWithIdAndInformationOnMassImportError,
-  SessionWithMissingAbortReasonError,
-  SessionWithoutStartedCertificationError,
   SIECLE_ERRORS,
   SiecleXmlImportError,
   StageModificationForbiddenForLinkedTargetProfileError,
